+++ source/backend/Risk.API/Services/IMsjService.cs
@@ -31,7 +31,7 @@ namespace Risk.API.Services
         Respuesta<Pagina<Mensaje>> ListarMensajesPendientes(PaginaParametros paginaParametros = null);
         Respuesta<Pagina<Correo>> ListarCorreosPendientes(PaginaParametros paginaParametros = null);
         Respuesta<Pagina<Notificacion>> ListarNotificacionesPendientes(PaginaParametros paginaParametros = null);
-        Respuesta<Dato> CambiarEstadoMensajeria(TipoMensajeria tipoMensajeria, int idMensajeria, EstadoMensajeria estado, string respuestaEnvio);
+        Respuesta<Dato> CambiarEstadoMensajeria(TipoMensajeria tipoMensajeria, long idMensajeria, EstadoMensajeria estado, string respuestaEnvio);
         Respuesta<Dato> ActivarDesactivarMensajeria(TipoMensajeria tipoMensajeria, bool estado);
     }
 }
